requirements.txt
flask
azure-ai-textanalytics
azure-core
python-dotenv
gunicorn
flask-session